Return the http_status_code specified here back to the caller
What this actually does is it raises a generic exception for all http
statuses that are not 200.

The Lambda + API Gateway integration works such that exceptions trigger
API Gateway to serialize the exception string into an `errorMessage`
key. This key can in turn be used to determine what the actual http
status code is supposed to be, and return that status code back to the
caller.

This is unfortunately not very intuitive, but the best that I have to
work with for now.