Encode the GitHub username inside the JSON Web Token
The reasoning for this is a combination of GitHub's API paths
(`/users/<login>` vs `/user`) and ember data.

If this login information were not inside the JWT, GitHub would need to
be queried via the `/user` endpoint. This is well and good until ember
data comes into play and likes pluralization to be a certain way.

What this came down to was either modify the ember data model (update
pluralization, among other things), or add the login field to the
JWT.

This seemed like the easier and more maintainable approach.